gateway-docker/requirements.txt
fastapi==0.112.2
pydantic==2.8.2
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.30.6
python-multipart